+++ apps/events/message.js
@@ -14,7 +14,7 @@ Bot.on?.("message", async(e) => {
   if (e.message_type == "group") {
     // 关闭撤回停止存储
     if (Config.getGroup(e.group_id).groupRecall) {
-      // logger.debug(`[Yenai-Plugin]存储群消息${e.group_id}=>${e.message_id}`)
+      // logger.debug(`[yenai-copy-plugin]存储群消息${e.group_id}=>${e.message_id}`)
       // 写入
       await redis.set(
         `notice:messageGroup:${e.message_id}`,
@@ -25,7 +25,7 @@ Bot.on?.("message", async(e) => {
   } else if (e.message_type == "private") {
     // 关闭撤回停止存储
     if (Config.whole.PrivateRecall) {
-      // logger.debug(`[Yenai-Plugin]存储私聊消息${e.user_id}=>${e.message_id}`)
+      // logger.debug(`[yenai-copy-plugin]存储私聊消息${e.user_id}=>${e.message_id}`)
       // 写入
       await redis.set(
         `notice:messagePrivate:${e.message_id}`,
@@ -42,7 +42,7 @@ Bot.on?.("message", async(e) => {
     e.message_type === "group"
   ) {
     if (!Config.getGroup(e.group_id).flashPhoto) return false
-    logger.info("[Yenai-Plugin]群聊闪照")
+    logger.info("[yenai-copy-plugin]群聊闪照")
     msg = [
       segment.image(`https://p.qlogo.cn/gh/${e.group_id}/${e.group_id}/100`),
       `[消息(${e.self_id}) - 闪照消息]\n`,
@@ -57,7 +57,7 @@ Bot.on?.("message", async(e) => {
     e.message_type === "discuss" &&
     Config.whole.flashPhoto
   ) {
-    logger.info("[Yenai-Plugin]讨论组闪照")
+    logger.info("[yenai-copy-plugin]讨论组闪照")
     msg = [
       segment.image(`https://q1.qlogo.cn/g?b=qq&s=100&nk=${e.user_id}`),
       `[消息(${e.self_id}) - 闪照消息]\n`,
@@ -72,7 +72,7 @@ Bot.on?.("message", async(e) => {
     e.message_type === "private" &&
     Config.whole.flashPhoto
   ) {
-    logger.info("[Yenai-Plugin]好友闪照")
+    logger.info("[yenai-copy-plugin]好友闪照")
     msg = [
       segment.image(`https://q1.qlogo.cn/g?b=qq&s=100&nk=${e.user_id}`),
       `[消息(${e.self_id}) - 闪照消息]\n`,
@@ -89,7 +89,7 @@ Bot.on?.("message", async(e) => {
       forwardMsg = arr.msg
       e.message = arr.type
     }
-    logger.info("[Yenai-Plugin]好友消息")
+    logger.info("[yenai-copy-plugin]好友消息")
     msg = [
       segment.image(`https://q1.qlogo.cn/g?b=qq&s=100&nk=${e.user_id}`),
       `[消息(${e.self_id}) - 好友消息]\n`,
@@ -116,7 +116,7 @@ Bot.on?.("message", async(e) => {
       forwardMsg = arr.msg
       e.message = arr.type
     }
-    logger.info("[Yenai-Plugin]群临时消息")
+    logger.info("[yenai-copy-plugin]群临时消息")
     // 发送的消息
     msg = [
       segment.image(`https://q1.qlogo.cn/g?b=qq&s=100&nk=${e.user_id}`),
@@ -143,7 +143,7 @@ Bot.on?.("message", async(e) => {
       forwardMsg = arr.msg
       e.message = arr.type
     }
-    logger.info("[Yenai-Plugin]群聊消息")
+    logger.info("[yenai-copy-plugin]群聊消息")
     msg = [
       segment.image(`https://p.qlogo.cn/gh/${e.group_id}/${e.group_id}/100`),
       `[消息(${e.self_id}) - 群聊消息]\n`,
@@ -156,7 +156,7 @@ Bot.on?.("message", async(e) => {
     ]
   } else if (e.message_type === "discuss") {
     if (!Config.getGroup(e.group_id).groupMessage) return false
-    logger.info("[Yenai-Plugin]讨论组消息")
+    logger.info("[yenai-copy-plugin]讨论组消息")
     msg = [
       segment.image(`https://q1.qlogo.cn/g?b=qq&s=100&nk=${e.user_id}`),
       `[消息(${e.self_id}) - 群聊消息]\n`,