+++ components/Config.js
@@ -5,7 +5,7 @@ import YamlReader from "./YamlReader.js"
 import cfg from "../../../lib/config/config.js"
 
 const Path = process.cwd()
-const Plugin_Name = "yenai-plugin"
+const Plugin_Name = "yenai-copy-plugin"
 const Plugin_Path = `${Path}/plugins/${Plugin_Name}`
 class Config {
   constructor() {
@@ -165,7 +165,7 @@ class Config {
     watcher.on("change", path => {
       delete this.config[key]
       if (typeof Bot == "undefined") return
-      logger.mark(`[Yenai-Plugin][修改配置文件][${type}][${name}]`)
+      logger.mark(`[yenai-copy-plugin][修改配置文件][${type}][${name}]`)
       if (this[`change_${name}`]) {
         this[`change_${name}`]()
       }