+++ config/system/sese_system.js
@@ -1,7 +1,7 @@
 /** 涩涩帮助 */
 export const helpCfg = {
   title: "椰奶涩涩帮助",
-  subTitle: "Yunzai-Bot & Yenai-Plugin",
+  subTitle: "Yunzai-Bot & yenai-copy-plugin",
   columnCount: 3,
   colWidth: 265,
   theme: "all",