+++ model/State/FastFetch.js
@@ -6,9 +6,9 @@ import { Config } from "../../components/index.js"
  */
 export default async function getFastFetch(e) {
   if (!isFeatureVisible(e.isPro)) return ""
-  let ret = await execSync("bash plugins/yenai-plugin/resources/state/state.sh")
+  let ret = await execSync("bash plugins/yenai-copy-plugin/resources/state/state.sh")
   if (ret.error) {
-    logger.error(`[Yenai-Plugin][状态]Error FastFetch 请检查是否使用git bash启动Yunzai-bot，错误信息：${ret.stderr}`)
+    logger.error(`[yenai-copy-plugin][状态]Error FastFetch 请检查是否使用git bash启动Yunzai-bot，错误信息：${ret.stderr}`)
     return ""
   }
   return ret.stdout.trim()