+++ model/State/GPU.js
@@ -22,7 +22,7 @@ export default async function getGPU() {
       item.memoryUsed && item.memoryFree && item.utilizationGpu
     )
     if (!graphics) {
-      logger.warn("[Yenai-plugin][state]状态GPU数据异常：\n", controllers)
+      logger.warn("[yenai-copy-plugin][state]状态GPU数据异常：\n", controllers)
       return false
     }
     let {
@@ -41,7 +41,7 @@ export default async function getGPU() {
       ]
     }
   } catch (e) {
-    logger.warn("[Yenai-Plugin][State] 获取GPU失败")
+    logger.warn("[yenai-copy-plugin][State] 获取GPU失败")
     return false
   }
 }