+++ model/State/style.js
@@ -29,13 +29,13 @@ export async function getBackground() {
     const fileSizeInBytes = buffer.byteLength
     const fileSizeInKB = (fileSizeInBytes / 1024).toFixed(2)
 
-    logger.info(`[Yenai-Plugin][状态]背景图片请求成功 ${fileSizeInKB}KB ${elapsedTime}ms`)
+    logger.info(`[yenai-copy-plugin][状态]背景图片请求成功 ${fileSizeInKB}KB ${elapsedTime}ms`)
 
     const buffBase64 = arrayBufferToBase64(buffer)
     return `data:image/jpeg;base64,${buffBase64}`
   } catch (err) {
     const bg = getDefaultBackdrop(backdropDefault)
-    backdrop && logger.warn(`[Yenai-Plugin][状态]背景图请求失败，使用默认背景图 “${bg}” ，错误原因: ${err.message}`)
+    backdrop && logger.warn(`[yenai-copy-plugin][状态]背景图请求失败，使用默认背景图 “${bg}” ，错误原因: ${err.message}`)
     return bg
   } finally {
     clearTimeout()
@@ -46,11 +46,11 @@ function arrayBufferToBase64(arrayBuffer) {
   return Buffer.from(arrayBuffer).toString("base64")
 }
 function getDefaultBackdrop(backdropDefault) {
-  const Plugin_Path = "../../../../../plugins/yenai-plugin"
-  const Bg_Path = "./plugins/yenai-plugin/resources/state/img/bg"
+  const Plugin_Path = "../../../../../plugins/yenai-copy-plugin"
+  const Bg_Path = "./plugins/yenai-copy-plugin/resources/state/img/bg"
   if (backdropDefault === "random") {
     backdropDefault = _.sample(fs.readdirSync(Bg_Path))
-    logger.debug(`[Yenai-Plugin][状态]使用随机背景图 “${backdropDefault}”`)
+    logger.debug(`[yenai-copy-plugin][状态]使用随机背景图 “${backdropDefault}”`)
   }
   return `${Plugin_Path}/resources/state/img/bg/${backdropDefault}`
 }